+++ src/__init__.py
@@ -1,6 +1,6 @@
 # coding=utf-8
 
-__version__ = '0.1.1'
+__version__ = '0.1.2'
 
 __title__ = "AviationWeather"
 __description__ = "A utility for both storing aviation weather into SQL " \