+++ CHANGELOG.rst
@@ -12,7 +12,18 @@ Initial commit.
 Version 0.1.1
 -------------
 Added README.txt
+
 Added sqlalchemy-utils to package requirements
+
 Fixed path errors in tests
+
 Fixed sample data errors in tests.
 
+-------------
+Version 0.1.2
+-------------
+Added exception handling in case of Database access error.
+
+Removed empty tests.
+
+Added function to delete old weather data.